endpoint: move proxystats functions to `proxystats.go`
Signed-off by: Ian Vernon <ian@cilium.io>
