+++ pkg/endpoint/proxystats.go
@@ -18,7 +18,83 @@ import (
 	"sort"
 
 	"github.com/cilium/cilium/api/v1/models"
+	"github.com/cilium/cilium/pkg/logging/logfields"
+	"github.com/cilium/cilium/pkg/metrics"
+	"github.com/cilium/cilium/pkg/policy"
+	"github.com/cilium/cilium/pkg/proxy/accesslog"
 )
+
+// getProxyStatisticsLocked gets the ProxyStatistics for the flows with the
+// given characteristics, or adds a new one and returns it.
+// Must be called with e.proxyStatisticsMutex held.
+func (e *Endpoint) getProxyStatisticsLocked(key string, l7Protocol string, port uint16, ingress bool) *models.ProxyStatistics {
+	if e.proxyStatistics == nil {
+		e.proxyStatistics = make(map[string]*models.ProxyStatistics)
+	}
+
+	proxyStats, ok := e.proxyStatistics[key]
+	if !ok {
+		var location string
+		if ingress {
+			location = models.ProxyStatisticsLocationIngress
+		} else {
+			location = models.ProxyStatisticsLocationEgress
+		}
+		proxyStats = &models.ProxyStatistics{
+			Location: location,
+			Port:     int64(port),
+			Protocol: l7Protocol,
+			Statistics: &models.RequestResponseStatistics{
+				Requests:  &models.MessageForwardingStatistics{},
+				Responses: &models.MessageForwardingStatistics{},
+			},
+		}
+
+		e.proxyStatistics[key] = proxyStats
+	}
+
+	return proxyStats
+}
+
+// UpdateProxyStatistics updates the Endpoint's proxy  statistics to account
+// for a new observed flow with the given characteristics.
+func (e *Endpoint) UpdateProxyStatistics(l4Protocol string, port uint16, ingress, request bool, verdict accesslog.FlowVerdict) {
+	e.proxyStatisticsMutex.Lock()
+	defer e.proxyStatisticsMutex.Unlock()
+
+	key := policy.ProxyID(e.ID, ingress, l4Protocol, port)
+	proxyStats, ok := e.proxyStatistics[key]
+	if !ok {
+		e.getLogger().WithField(logfields.L4PolicyID, key).Warn("Proxy stats not found when updating")
+		return
+	}
+
+	var stats *models.MessageForwardingStatistics
+	if request {
+		stats = proxyStats.Statistics.Requests
+	} else {
+		stats = proxyStats.Statistics.Responses
+	}
+
+	stats.Received++
+	metrics.ProxyReceived.Inc()
+	metrics.ProxyPolicyL7Total.WithLabelValues("received").Inc()
+
+	switch verdict {
+	case accesslog.VerdictForwarded:
+		stats.Forwarded++
+		metrics.ProxyForwarded.Inc()
+		metrics.ProxyPolicyL7Total.WithLabelValues("forwarded").Inc()
+	case accesslog.VerdictDenied:
+		stats.Denied++
+		metrics.ProxyDenied.Inc()
+		metrics.ProxyPolicyL7Total.WithLabelValues("denied").Inc()
+	case accesslog.VerdictError:
+		stats.Error++
+		metrics.ProxyParseErrors.Inc()
+		metrics.ProxyPolicyL7Total.WithLabelValues("parse_errors").Inc()
+	}
+}
 
 // sortProxyStats sorts the given slice of ProxyStatistics.
 func sortProxyStats(proxyStats []*models.ProxyStatistics) {
